pyproject.toml
[project]
name = "pgsql-train"
version = "0.1.0"
description = "PostgreSQL training project with Docker"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
